move game logic into +page.server.ts
